Coming back from a long time off. Trying to figure out the state of everything. This commit message is not that specific...
What it looks like I tried to do, is explictly map in:
- page ledger. The idea is that this ledger is will be used to track
  the physical page frames.

I noticed that I kept a pointer after the ledger that I increment and
mark in use when I allocate new page tables. Page tables have no need for
a virtual address, thus should be ignored.

I going to run with this concept and convert the memory_pool_ptr to
a pointer to a region of free kernel virtual memory addresses.