Added code to handle simple lists
Started to implement buddy code; some of it is in. However it is bad and
probably not working. I stopped working on the heap code because I need to
allocate larger contiguous blocks of physical memory for the heap. So I need
a better physical memory allocater. Enter the idea for the buddy system.